Fix markdownProcessor case
